commit built javascript files (for tests)
